EZP-24361: Extracted fields rendering to dedicated Twig extension
